requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   openai
asgiref==3.11.0
    # via
    #   django
    #   django-cors-headers
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.14.3
    # via django-unicorn
cachetools==6.2.4
    # via django-unicorn
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
concurrent-log-handler==0.9.28
    # via business-suite (pyproject.toml)
contourpy==1.3.3
    # via matplotlib
crispy-bootstrap5==2025.6
    # via business-suite (pyproject.toml)
cycler==0.12.1
    # via matplotlib
decorator==5.2.1
    # via django-unicorn
distro==1.9.0
    # via openai
django==6.0
    # via
    #   business-suite (pyproject.toml)
    #   crispy-bootstrap5
    #   django-auditlog
    #   django-cors-headers
    #   django-crispy-forms
    #   django-dbbackup
    #   django-debug-toolbar
    #   django-storages
    #   django-unicorn
    #   django-user-agents
    #   django-waffle
    #   django-widget-tweaks
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-auditlog==3.4.1
    # via business-suite (pyproject.toml)
django-bootstrap-sidebar==1.2.7
    # via business-suite (pyproject.toml)
django-cleanup==9.0.0
    # via business-suite (pyproject.toml)
django-cors-headers==4.9.0
    # via business-suite (pyproject.toml)
django-crispy-forms==2.5
    # via crispy-bootstrap5
django-dbbackup==5.1.0
    # via business-suite (pyproject.toml)
django-debug-toolbar==6.1.0
    # via business-suite (pyproject.toml)
django-nested-admin==4.1.6
    # via business-suite (pyproject.toml)
django-storages==1.14.6
    # via business-suite (pyproject.toml)
django-unicorn==0.62.0
    # via business-suite (pyproject.toml)
django-user-agents==0.4.0
    # via business-suite (pyproject.toml)
django-waffle==5.0.0
    # via business-suite (pyproject.toml)
django-widget-tweaks==1.5.1
    # via business-suite (pyproject.toml)
djangorestframework==3.16.1
    # via
    #   business-suite (pyproject.toml)
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-camel-case==1.4.2
    # via business-suite (pyproject.toml)
djangorestframework-simplejwt==5.3.1
    # via business-suite (pyproject.toml)
docx-mailmerge==0.5.0
    # via business-suite (pyproject.toml)
drf-spectacular==0.29.0
    # via business-suite (pyproject.toml)
dropbox==12.0.2
    # via business-suite (pyproject.toml)
et-xmlfile==2.0.0
    # via openpyxl
fonttools==4.61.1
    # via matplotlib
gunicorn==23.0.0
    # via business-suite (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
huey==2.5.3
    # via business-suite (pyproject.toml)
hypothesis==6.148.11
    # via business-suite (pyproject.toml)
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
imageio==2.37.2
    # via
    #   passporteye
    #   scikit-image
inflection==0.5.1
    # via drf-spectacular
jiter==0.12.0
    # via openai
joblib==1.5.3
    # via scikit-learn
jsonschema==4.26.0
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
kiwisolver==1.4.9
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
lxml==6.0.2
    # via
    #   docx-mailmerge
    #   python-docx
matplotlib==3.10.8
    # via
    #   passporteye
    #   seaborn
networkx==3.6.1
    # via scikit-image
numpy==2.4.0
    # via
    #   contourpy
    #   imageio
    #   matplotlib
    #   pandas
    #   passporteye
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   seaborn
    #   tifffile
openai==2.14.0
    # via business-suite (pyproject.toml)
openpyxl==3.1.5
    # via business-suite (pyproject.toml)
orjson==3.11.5
    # via django-unicorn
packaging==25.0
    # via
    #   gunicorn
    #   lazy-loader
    #   matplotlib
    #   pytesseract
    #   scikit-image
pandas==2.3.3
    # via seaborn
passporteye==2.2.2
    # via business-suite (pyproject.toml)
pdf2image==1.17.0
    # via business-suite (pyproject.toml)
pdfminer==20191125
    # via passporteye
peewee==3.19.0
    # via business-suite (pyproject.toml)
pillow==12.1.0
    # via
    #   business-suite (pyproject.toml)
    #   imageio
    #   matplotlib
    #   pdf2image
    #   pytesseract
    #   scikit-image
ply==3.11
    # via stone
portalocker==3.2.0
    # via concurrent-log-handler
psycopg2-binary==2.9.11
    # via business-suite (pyproject.toml)
pycryptodome==3.23.0
    # via pdfminer
pydantic==2.12.5
    # via openai
pydantic-core==2.41.5
    # via pydantic
pyjwt==2.11.0
    # via djangorestframework-simplejwt
pymemcache==4.0.0
    # via business-suite (pyproject.toml)
pyparsing==3.3.1
    # via matplotlib
pypdf==6.5.0
    # via business-suite (pyproject.toml)
pytesseract==0.3.13
    # via passporteye
python-dateutil==2.9.0.post0
    # via
    #   django-auditlog
    #   matplotlib
    #   pandas
python-docx==1.2.0
    # via business-suite (pyproject.toml)
python-dotenv==1.2.1
    # via business-suite (pyproject.toml)
python-logging-loki==0.3.1
    # via business-suite (pyproject.toml)
python-monkey-business==1.1.0
    # via django-nested-admin
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   business-suite (pyproject.toml)
    #   drf-spectacular
rapidfuzz==3.14.3
    # via business-suite (pyproject.toml)
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   business-suite (pyproject.toml)
    #   dropbox
    #   python-logging-loki
rfc3339==6.2
    # via python-logging-loki
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
scikit-image==0.26.0
    # via passporteye
scikit-learn==1.8.0
    # via passporteye
scipy==1.16.3
    # via
    #   passporteye
    #   scikit-image
    #   scikit-learn
seaborn==0.13.2
    # via business-suite (pyproject.toml)
setuptools==80.9.0
    # via business-suite (pyproject.toml)
shortuuid==1.0.13
    # via django-unicorn
six==1.17.0
    # via
    #   dropbox
    #   python-dateutil
    #   stone
sniffio==1.3.1
    # via openai
sortedcontainers==2.4.0
    # via hypothesis
soupsieve==2.8.1
    # via beautifulsoup4
sqlparse==0.5.5
    # via
    #   django
    #   django-debug-toolbar
stone==3.3.1
    # via dropbox
threadpoolctl==3.6.0
    # via scikit-learn
tifffile==2025.12.20
    # via scikit-image
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   openai
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
ua-parser==1.0.1
    # via
    #   business-suite (pyproject.toml)
    #   user-agents
ua-parser-builtins==202601
    # via ua-parser
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.6.2
    # via requests
user-agents==2.2.0
    # via
    #   business-suite (pyproject.toml)
    #   django-user-agents
whitenoise==6.11.0
    # via business-suite (pyproject.toml)
